redisigned request_log app structure
