+++ snort3_monitor/monitor/models.py
@@ -1,4 +1,5 @@
 from django.db import models
+
 from rule.models import Rule
 
 
@@ -11,12 +12,3 @@ class Event(models.Model):
     dst_port = models.IntegerField(null=True, blank=True)
     proto = models.CharField(max_length=128)
     mark_as_deleted = models.BooleanField(default=False)
-
-
-class Request(models.Model):
-    user_addr = models.CharField(max_length=50)
-    http_method = models.CharField(max_length=20)
-    timestamp = models.DateTimeField(auto_now_add=True)
-    endpoint = models.CharField(max_length=128)
-    response = models.IntegerField()
-    request_data = models.JSONField()
