+++ snort3_monitor/monitor/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from monitor.views import EventListUpdate, EventCountList
+
+urlpatterns = [
+    path("", EventListUpdate.as_view(), name='event-list-update'),
+    path("count/", EventCountList.as_view(), name='event-count-list'),
+]
