+++ snort3_monitor/monitor/views.py
@@ -1,25 +1,15 @@
-import logging
 from datetime import datetime, timedelta
 
 from django.db.models import Count, Value
 from django.db.models.functions import Concat
 from django.db.models import F
 from django.db.models.query import QuerySet
-from django.utils.timezone import make_aware, utc
 from rest_framework import generics, status
 from rest_framework.exceptions import ValidationError
 from rest_framework.response import Response
 
-from .models import Event, Request
-from .serializers import EventSerializer, EventCountAddressSerializer, EventCountRuleSerializer, RequestSerializer
-
-
-logger = logging.getLogger('monitor')
-formatter = logging.Formatter('%(name)s -> %(levelname)s : %(message)s')
-handler = logging.StreamHandler()
-handler.setLevel(logging.INFO)
-handler.setFormatter(formatter)
-logger.addHandler(handler)
+from .models import Event
+from .serializers import EventSerializer, EventCountAddressSerializer, EventCountRuleSerializer
 
 
 class EventListUpdate(generics.UpdateAPIView, generics.ListAPIView):
@@ -58,54 +48,6 @@ class EventListUpdate(generics.UpdateAPIView, generics.ListAPIView):
         queryset = self.get_queryset()
         queryset.update(mark_as_deleted=True)
         return Response({"message": "All events are marked as deleted."}, status=status.HTTP_200_OK)
-
-
-class RequestList(generics.ListAPIView):
-    queryset = Request.objects.all()
-    serializer_class = RequestSerializer
-
-    def get_queryset(self) -> QuerySet:
-        """Perform filtering and ordering data
-
-        Client can use only allowed_params in query.
-        """
-        queryset = super().get_queryset()
-
-        allowed_params = ['period_start', 'period_stop']
-        params = [key for key in self.request.query_params]
-        validate_params(params, allowed_params)
-
-        # checks if all params are included
-        period_start = self.request.query_params.get('period_start')
-        period_stop = self.request.query_params.get('period_stop')
-        if not (period_start and period_stop):
-            raise ValidationError({"error": "You should define 'period_start' and 'period_stop'."})
-
-        # checks if format is proper
-        period_start = self.validate_date(period_start)
-        period_stop = self.validate_date(period_stop) + timedelta(days=1)
-
-        # checks if period is less than week
-        if period_stop - period_start > timedelta(days=7):
-            raise ValidationError({"error": "The range has to be less than a week"})
-
-        queryset = queryset.filter(timestamp__gte=period_start, timestamp__lte=period_stop)
-        return queryset.order_by('id')
-
-    @staticmethod
-    def validate_date(date: str) -> datetime:
-        """Validate date format
-
-        Only formats(list) are allowed in query.
-        """
-        formats = ["%Y-%m-%d %H:%M:%S", "%Y-%m-%d %H:%M", "%Y-%m-%d %H", "%Y-%m-%d"]
-        for frmt in formats:
-            try:
-                date = datetime.strptime(date, frmt)
-                return make_aware(date, utc)
-            except ValueError:
-                pass
-        raise ValidationError({"error": "Use format YYYY-MM-DD HH:MM:SS (you can skip SS, MM, HH)"})
 
 
 class EventCountList(generics.ListAPIView):
