+++ snort3_monitor/request_log/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class RequestLogConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'request_log'
