+++ snort3_monitor/request_log/serializers.py
@@ -1,0 +1,9 @@
+from rest_framework import serializers
+
+from request_log.models import Request
+
+
+class RequestSerializer(serializers.ModelSerializer):
+    class Meta:
+        model = Request
+        fields = '__all__'
