+++ snort3_monitor/request_log/urls.py
@@ -1,0 +1,8 @@
+from django.urls import path
+
+from request_log.views import RequestList
+
+
+urlpatterns = [
+    path("", RequestList.as_view(), name='request-list'),
+]
